requirements.txt
openai
discord.py
dotenv
python-dotenv
